refactor: nightMate
