+++ ZANI/Data/Repositories/DefaultTeamRepository.swift
@@ -10,7 +10,7 @@ import Foundation
 
 final class DefaultTeamRepository: BaseService, TeamRepository {
   
-  func requestTeamList(data: RequestTeamListDTO, completion: @escaping (NetworkResult<Any>) -> Void) {
+  public func requestTeamList(data: RequestTeamListDTO, completion: @escaping (NetworkResult<Any>) -> Void) {
     AFManager.request(
       TeamRouter.requestTeamList(
         keyword: data.keyword,
@@ -34,7 +34,7 @@ final class DefaultTeamRepository: BaseService, TeamRepository {
     }
   }
   
-  func requestCreateTeam(data: RequestCreateTeamDTO, completion: @escaping (NetworkResult<Any>) -> Void) {
+  public func requestCreateTeam(data: RequestCreateTeamDTO, completion: @escaping (NetworkResult<Any>) -> Void) {
     AFManager.request(
       TeamRouter.requestCreateTeam(
         title: data.title,
@@ -51,6 +51,28 @@ final class DefaultTeamRepository: BaseService, TeamRepository {
         guard let statusCode = response.response?.statusCode else { return }
         guard let data = response.data else { return }
         let networkResult = self.judgeStatus(by: statusCode, data, type: RecruitmentTeamData.self)
+        completion(networkResult)
+        
+      case .failure(let err):
+        print(err.localizedDescription)
+      }
+    }
+  }
+  
+  public func requestChatHistory(
+    teamId: Int,
+    page: Int,
+    size: Int,
+    completion: @escaping (NetworkResult<Any>) -> Void
+  ) {
+    AFManager.request(
+      ChattingRouter.requestChattingHistory(teamId: teamId, page: page, size: size)
+    ).responseData { response in
+      switch response.result {
+      case .success:
+        guard let statusCode = response.response?.statusCode else { return }
+        guard let data = response.data else { return }
+        let networkResult = self.judgeStatus(by: statusCode, data, type: ChatDTO.self)
         completion(networkResult)
         
       case .failure(let err):
